pyproject.toml
[project]
name = "corpus-rag"
version = "0.1.0"
description = "Single-user RAG framework for personal archives. Local-first, MCP-native."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "corpus contributors" }]
keywords = ["rag", "retrieval", "mcp", "embeddings", "personal-archive"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Text Processing :: Indexing",
]
dependencies = [
    # Pin compatible ranges, not floors — Voyage already shipped breaking-ish
    # changes mid-0.x. Same for the other 0.x SDKs.
    "voyageai>=0.3.7,<0.4",
    "sqlite-vec>=0.1.9,<0.2",
    "mcp>=1.27.1,<2.0",
    "pydantic>=2.13.4,<3.0",
    "python-dotenv>=1.2.2,<2.0",
    "tomli>=2.0; python_version < '3.11'",  # 3.12 has tomllib stdlib
]

[project.optional-dependencies]
# Optional: local cross-encoder re-ranker. ~2GB of deps (torch + transformers
# + sentence-transformers); only install if you actually want re-ranking.
reranker = ["sentence-transformers>=5.5.0,<6.0"]
# Optional: Claude Haiku per-doc summaries.
summarizer = ["anthropic>=0.50,<1.0"]
pdf = ["pypdf>=6.11.0,<7.0"]
html = ["trafilatura>=2.0.0,<3.0"]
gemini = ["google-genai>=2.2.0,<3.0"]
all = [
    "sentence-transformers>=5.5.0,<6.0",
    "anthropic>=0.50,<1.0",
    "pypdf>=6.11.0,<7.0",
    "trafilatura>=2.0.0,<3.0",
    "google-genai>=2.2.0,<3.0",
]

[project.scripts]
corpus-init = "corpus.cli.init:main"
corpus-ingest = "corpus.cli.ingest:main"
corpus-query = "corpus.cli.query:main"
corpus-eval = "corpus.cli.eval:main"
corpus-mcp = "corpus.mcp_server:main"
corpus-summarize = "corpus.cli.summarize:main"
corpus-benchmark = "corpus.cli.benchmark:main"
corpus-list = "corpus.cli.list_sources:main"
corpus-reset = "corpus.cli.reset:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/corpus"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.12",
    "mypy>=2.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501", "RUF001", "RUF003", "SIM108", "UP047", "SIM103"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unreachable = true

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
